pyproject.toml
[project]
name = "cicloapi"
version = "0.2"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "M. Herrero", email = "mherrero@bsc.es" }
]
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi==0.115.6",
    "geojson>=3.2.0",
    "geopandas>=0.14",
    "h3>=4.2.0",
    "haversine>=2.9.0",
    "matplotlib>=3.10.0",
    "opencv-python>=4.11.0.86",
    "osmnx>=1.9.4",
    "passlib[bcrypt]==1.7.4",
    "pathlib>=1.0.1",
    "pyjwt==2.10.1",
    "python-dotenv>=1.0.1",
    "python-igraph>=0.11.8",
    "python-multipart==0.0.20",
    "rasterio>=1.4.3",
    "tobler>=0.12.1",
    "tqdm>=4.67.1",
    "uvicorn==0.34.0",
    "watermark>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
packages = ["cicloapi"]
package-dir = {"" = "src"}

[project.scripts]
post_install = "cicloapi.install_hook:generate_env_variable"
cicloapi = "cicloapi.__init__:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cicloapi*"]

[tool.setuptools.package_data]
"cicloapi" = [
    "src/cicloapi/data/*.json",
    "src/cicloapi/*.py",
]

